requirements.txt
Django==1.6.1
MySQL-python==1.2.5
South==0.8.4
django-braces==1.3.1
httplib2==0.8
ipdb==0.8
ipython==1.2.0
pyreadline==2.0
six==1.5.2
twilio==3.6.5